backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlalchemy
psycopg[binary]
asyncpg
alembic
httpx
python-dotenv
websockets
cryptography
snowflake-connector-python
